Add Check Permission in Controllers, add useDynamicForm function in ts

+++ Modules/Hr/app/Http/Controllers/Organization/PositionController.php
@@ -10,6 +10,7 @@ use Modules\Hr\Models\Department;
 use Modules\Hr\Services\PositionService;
 use Modules\Hr\Http\Requests\PositionRequest;
 use App\Http\Controllers\TransactionController;
+use Illuminate\Support\Facades\Gate;
 
 class PositionController extends TransactionController
 {
@@ -22,6 +23,8 @@ class PositionController extends TransactionController
      */
     public function index()
     {
+        abort_if(Gate::denies('position_access'), 403);
+
         return Inertia::render($this->path . 'PositionsListComponent', [
             'positions'     => Position::with(['department', 'division', 'section'])->filter(request()->query() ?? [])->paginate(10),
             'divisions'     => Division::pluck('name', 'id'),
@@ -35,6 +38,8 @@ class PositionController extends TransactionController
      */
     public function store(PositionRequest $request)
     {
+        abort_if(Gate::denies('position_create'), 403);
+
         return $this->withTransaction(function () use ($request) {
             $this->positionService->store($request->validated());
             return redirect()->route('hr.organization.positions.index');
@@ -46,6 +51,8 @@ class PositionController extends TransactionController
      */
     public function update(PositionRequest $request, Position $position)
     {
+        abort_if(Gate::denies('position_edit'), 403);
+
         return $this->withTransaction(function () use ($request, $position) {
             $this->positionService->update($position, $request->validated());
             return redirect()->route('hr.organization.positions.index');
@@ -57,6 +64,8 @@ class PositionController extends TransactionController
      */
     public function destroy(Position $position)
     {
+        abort_if(Gate::denies('position_delete'), 403);
+
         return $this->withTransaction(function () use ($position) {
             $this->positionService->destroy($position);
             return redirect()->route('hr.organization.positions.index');
